repo-tools: update type-deps to analyze all entry points
