+++ packages/repo-tools/src/commands/type-deps/type-deps.ts
@@ -20,6 +20,7 @@ import { resolve as resolvePath } from 'path';
 // eslint-disable-next-line @backstage/no-undeclared-imports
 import chalk from 'chalk';
 import { getPackages, Package } from '@manypkg/get-packages';
+import { getPackageExportNames } from '../../lib/entryPoints';
 
 export default async () => {
   const { packages } = await getPackages(resolvePath('.'));
@@ -96,11 +97,15 @@ function findAllDeps(declSrc: string) {
  * missing or incorrect in package.json
  */
 function checkTypes(pkg: Package) {
-  const typeDecl = fs.readFileSync(
-    resolvePath(pkg.dir, 'dist/index.d.ts'),
-    'utf8',
-  );
-  const allDeps = findAllDeps(typeDecl);
+  const entryPointNames = getPackageExportNames(pkg.packageJson) ?? ['index'];
+
+  const allDeps = entryPointNames.flatMap(name => {
+    const typeDecl = fs.readFileSync(
+      resolvePath(pkg.dir, `dist/${name}.d.ts`),
+      'utf8',
+    );
+    return findAllDeps(typeDecl);
+  });
   const deps = Array.from(new Set(allDeps));
 
   const errors = [];
